requirements.txt
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
boto3==1.18.16
botocore==1.21.16
cffi==1.14.6
channels==3.0.4
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2.5
django-storages==1.11.1
djangorestframework==3.12.4
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
jmespath==0.10.0
Pillow==8.3.1
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
s3transfer==0.5.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
Twisted==21.7.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
whitenoise==5.3.0
zope.interface==5.4.0
